backend/pyproject.toml
[project]
name = "flask-boss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-jwt-extended>=4.7.1",
    "flask-limiter>=3.12",
    "flask-sqlalchemy>=3.1.1",
    "pillow>=11.3.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "redis",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
